+++ cli/src/lib.rs
@@ -22,6 +22,7 @@ pub mod config;
 pub mod description_util;
 pub mod diff_util;
 pub mod formatter;
+pub mod git_util;
 pub mod graphlog;
 pub mod merge_tools;
 pub mod operation_templater;
